The funny band
This project is made in Scratch. It is a kind of a musical game. There are 3 actors in the game:  a pinguin, a monkey and a bear. They play several melodies. The task is to guess a singer. If the answer is correct, the bear starts to dance and the player gets 1 score.